Servo control through Timer/Counter 0
Initial version of the simplified servo control through timer/counter 0.

The previous code was done through the 16-bit timer/counter 1 and
implemented in software. This works (mostly) for other projects but
this very simple project has only one servo and it is suitable to waste
the entire Timer/Counter 0 for a single servo.

The control is done with Phase Correct PWM mode 1 through OCR0A
with 32 microsecond resolution providing about 3 degree resolution
for the servo itself.

No interrupt routine is being used. The servo is refreshed about 122
times
a second.

No data sheet was found for this specific servo, but another similarly
controlled servo on the same site has more details and says 330 Hz
refresh is acceptable.